requirements.txt
torch
matplotlib
wandb
neuraloperator